+++ src/hooks/useInfo.ts
@@ -1,0 +1,24 @@
+import { useState, useEffect } from 'react';
+import { invoke } from "@tauri-apps/api";
+
+export interface AppInfo {
+  version: string;
+  authors: string;
+  db_path: string;
+  total_tasks: number;
+}
+
+const useInfo = () => {
+  const [info, setInfo] = useState<AppInfo | null>(null);
+
+  useEffect(() => {
+    invoke("info").then((i) => {
+      console.log(i);
+      setInfo(i as AppInfo);
+    });
+  }, []);
+
+  return { info }
+}
+
+export default useInfo;