+++ src-tauri/src/main.rs
@@ -7,6 +7,8 @@ pub mod repositories;
 pub mod schema;
 pub mod utils;
 
+use std::process::Command;
+
 use models::Setting;
 use tauri::{command, SystemTray};
 
@@ -194,6 +196,53 @@ fn info(app_handle: tauri::AppHandle) -> Value {
     })
 }
 
+#[tauri::command]
+fn show_in_folder(path: String) {
+    #[cfg(target_os = "windows")]
+    {
+        Command::new("explorer")
+            .args(["/select,", &path]) // The comma after select is not a typo
+            .spawn()
+            .unwrap();
+    }
+
+    #[cfg(target_os = "linux")]
+    {
+        if path.contains(",") {
+            // see https://gitlab.freedesktop.org/dbus/dbus/-/issues/76
+            let new_path = match metadata(&path).unwrap().is_dir() {
+                true => path,
+                false => {
+                    let mut path2 = PathBuf::from(path);
+                    path2.pop();
+                    path2.into_os_string().into_string().unwrap()
+                }
+            };
+            Command::new("xdg-open").arg(&new_path).spawn().unwrap();
+        } else {
+            if let Ok(Fork::Child) = daemon(false, false) {
+                Command::new("dbus-send")
+                    .args([
+                        "--session",
+                        "--dest=org.freedesktop.FileManager1",
+                        "--type=method_call",
+                        "/org/freedesktop/FileManager1",
+                        "org.freedesktop.FileManager1.ShowItems",
+                        format!("array:string:\"file://{path}\"").as_str(),
+                        "string:\"\"",
+                    ])
+                    .spawn()
+                    .unwrap();
+            }
+        }
+    }
+
+    #[cfg(target_os = "macos")]
+    {
+        Command::new("open").args(["-R", &path]).spawn().unwrap();
+    }
+}
+
 fn main() {
     let system_tray = SystemTray::new();
     tauri::Builder::default()
@@ -218,6 +267,7 @@ fn main() {
             favourites,
             redmine_activities,
             info,
+            show_in_folder,
         ])
         .system_tray(system_tray)
         .run(tauri::generate_context!())